refactor: springboard > launchpad
