+++ packages/launchpad/cypress/.eslintrc.json
