+++ packages/launchpad/cypress/component/App.spec.ts
