+++ packages/launchpad/cypress/component/CloseButton.spec.ts
