+++ packages/launchpad/cypress/component/InstallDependencies.spec.ts
