+++ packages/launchpad/cypress/component/NewUserWelcome.spec.ts
