+++ packages/launchpad/cypress/component/TerminalCommand.spec.ts
