+++ packages/launchpad/cypress/integration/App.spec.ts
