+++ packages/launchpad/lib/gui.js
