+++ packages/launchpad/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@packages/springboard",
+  "name": "@packages/launchpad",
   "version": "0.0.0-development",
   "private": true,
   "main": "lib/gui.js",
@@ -11,8 +11,8 @@
     "cypress:open:ct": "cross-env TZ=America/New_York node ../../scripts/cypress open-ct --project .",
     "cypress:run": "cross-env TZ=America/New_York node ../../scripts/cypress run --project .",
     "cypress:run:ct": "cross-env TZ=America/New_York node ../../scripts/cypress run-ct --project .",
-    "dev": "SPRINGBOARD=1 yarn start-test 5050 cypress:open",
-    "postinstall": "echo '@packages/springboard needs: yarn build'",
+    "dev": "LAUNCHPAD=1 yarn start-test 5050 cypress:open",
+    "postinstall": "echo '@packages/launchpad needs: yarn build'",
     "start": "http-server -p 5050 dist",
     "watch": "yarn build --watch --progress"
   },
