+++ packages/launchpad/src/components/CloseButton.vue
