+++ packages/launchpad/src/components/DependencyItem.vue
