+++ packages/launchpad/src/components/InstallDependencies.vue
