+++ packages/launchpad/src/components/NewUserWelcome.vue
