+++ packages/launchpad/src/components/SelectWizard.vue
