+++ packages/launchpad/src/components/TerminalCommand.vue
