+++ packages/server/lib/gui/windows.ts
@@ -5,7 +5,7 @@ import { BrowserWindow } from 'electron'
 import Debug from 'debug'
 import cwd from '../cwd'
 import savedState from '../saved_state'
-const cyDesktopSpringboard = require('@packages/springboard')
+const cyDesktopLaunchpad = require('@packages/launchpad')
 const cyDesktopGui = require('@packages/desktop-gui')
 
 const debug = Debug('cypress:server:windows')
@@ -22,8 +22,8 @@ let recentlyCreatedWindow = false
 const getUrl = function (type) {
   switch (type) {
     case 'INDEX':
-      if (process.env.SPRINGBOARD) {
-        return cyDesktopSpringboard.getPathToIndex()
+      if (process.env.LAUNCHPAD) {
+        return cyDesktopLaunchpad.getPathToIndex()
       }
 
       return cyDesktopGui.getPathToIndex()
